fix(review): de-leak shooter composite (drop season-atlas gravity ingredient -- verdict survives leak-free: rho .289 ci_lo .0193, was inflated .335/.0615); SP gate excludes doubleheader dates (join fan-out inflated n/narrowed CI); finite-mask casts astype(float) both validators (object-dtype crash surface in pod loop)
Adversarial review verdict was BLOCK on 2 findings; both fixed, 30 tests green, harness re-run on real data.

+++ scripts/platformkit/intel_validation/claims_validator.py
@@ -147,7 +147,7 @@ def validate_claim(claim: dict[str, Any]) -> ClaimVerdict:
         # the producer applies (claims_factory._build_dim_claim), folded
         # into the same n_excluded count so the independent recompute stays
         # consistent with what a non-NaN-emitting producer would publish.
-        finite_mask = np.isfinite(recomputed["_value"])
+        finite_mask = np.isfinite(recomputed["_value"].astype(float))
         n_excluded += int((~finite_mask).sum())
         recomputed = recomputed[finite_mask]
         id_col = criteria["aggregate"]["group_by"]
